backend/requirements.txt
pydantic
auth0-python
python-dotenv
sqlalchemy
psycopg2-binary
httpx
fastapi
uvicorn
pydantic[email]
